extract download logic from exporters_controller into concern
